requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-filter==21.1
django-cors-headers==3.10.0
martor==1.6.44
gunicorn==23.0.0
python-dotenv=1.0.1